# Conflicts:
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.infra/src/main/java/nts/uk/ctx/at/shared/infra/entity/vacation/setting/nursingleave/KnlmtNursingLeaveSet.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.infra/src/main/java/nts/uk/ctx/at/shared/infra/repository/remainingnumber/nursingcareleavemanagement/info/JpaLeaveForCareInfoRepo.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.infra/src/main/java/nts/uk/ctx/at/shared/infra/repository/vacation/setting/nursingleave/JpaMaxPersonSettingGetMemento.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.infra/src/main/java/nts/uk/ctx/at/shared/infra/repository/vacation/setting/nursingleave/JpaNursingLeaveSettingGetMemento.java
#	nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.infra/src/main/java/nts/uk/ctx/at/shared/infra/repository/vacation/setting/nursingleave/JpaNursingLeaveSettingSetMemento.java
#	nts.uk/uk.at/at.screen/nts.uk.screen.at.app/src/main/java/nts/uk/screen/at/app/ktgwidget/find/OptionalWidgetKtgFinder.java

+++ nts.uk/uk.at/at.ctx/shared/nts.uk.ctx.at.shared.app/src/main/java/nts/uk/ctx/at/shared/app/command/vacation/setting/nursingleave/dto/NursingLeaveSettingDto.java
@@ -37,7 +37,7 @@ public class NursingLeaveSettingDto {
     private Integer startMonthDay;
 
     /** The nursing number leave day. */
-    private Integer nursingNumberLeaveDay;
+    private Double nursingNumberLeaveDay;
 
     /** The nursing number person. */
     private Integer nursingNumberPerson;
